requirements.txt
aiogram>=3.22.0
python-dotenv>=1.0.1
psutil>=5.9.8
pydantic>=2.11.7
